Start Selenium Server

Start Selenium Server

selenium(port = 4567L, version = "latest", chromever = "latest",
  geckover = "latest", iedrver = NULL, phantomver = "latest",
  verbose = TRUE)

Arguments

port

Port to run on

version

what version of Selenium Server to run. Default = "latest" which runs the most recent version. To see other version currently sourced run binman::list_versions("seleniumserver")

chromever

what version of Chrome driver to run. Default = "latest" which runs the most recent version. To see other version currently sourced run binman::list_versions("chromedriver"), A value of NULL excludes adding the chrome browser to Selenium Server.

geckover

what version of Gecko driver to run. Default = "latest" which runs the most recent version. To see other version currently sourced run binman::list_versions("geckodriver"), A value of NULL excludes adding the firefox browser to Selenium Server.

iedrver

what version of IEDriverServer to run. Default = "latest" which runs the most recent version. To see other version currently sourced run binman::list_versions("iedriverserver"), A value of NULL excludes adding the internet explorer browser to Selenium Server. NOTE this functionality is Windows OS only.

phantomver

what version of PhantomJS to run. Default = "latest" which runs the most recent version. To see other version currently sourced run binman::list_versions("phantomjs"), A value of NULL excludes adding the PhantomJS headless browser to Selenium Server.

verbose

If TRUE, include status messages (if any)

Value

Returns a list with named elements process, output, error and stop. process is the output from calling spawn_process output, error and stop are functions calling process_read, process_read with "stderr" pipe and process_kill respectively on process.

Examples

## Not run: ------------------------------------
# selServ <- selenium()
# selServ$output()
# selServ$stop()
## ---------------------------------------------